[FIR] Split function call CFG node into an entrance and exit
When non-local returns are enabled, jumps from within a lambda to
outside the lambda can leave the function exit as unreachable. This has
the unintended side effect of making the entire function call be
considered unreachable, because within the control flow graph, the
function call is only represented as a single node. There is no way to
indicate that the exit of the function is unreachable versus the
entrance.

By splitting the function call CFG node into an enter node and an exit
node, we can indicate that the non-local jump will make the function
exit node unreachable, but that the entrance remains reachable. This
will also allow the unreachable checker to only consider the function
call enter node when determining if the call is reachable.

^KT-68277 Fixed