[Swift export] Translate Kotlin Char to Foundation unichar
Currently, Swift export doesn't support Kotlin Char type (maps it to
an unsupported type).

We can't translate kotlin.Char to Swift.Character, because they have
different semantics: the former is a UTF-16 code unit, the latter is
a single extended grapheme cluster. So we can't fit some values of
Swift.Character into kotlin.Char.

Instead, there is a unichar type in the Apple Foundation framework.
It represents a UTF-16 element of NSString, which makes it similar to
kotlin.Char (and probably the closest option).

This commit implements translating kotlin.Char to Foundation.unichar
in Swift export:
* Handles kotlin.Char in the type provider
* Handles Foundation.unichar in the bridge generator

^KT-66862 Fixed