[FIR] Ignore plugin generated supertypes for `SUPERTYPE_NOT_INITIALIZED`
The checker checks source elements, but there are none.

Also note that for the `class WithImplicitAny` example to work properly
there must be a corresponding change on the IR side to replace the call
to the delegated constructor with the proper one: otherwise, it'll be a
call to the constructor of `Any`.

^KT-68375 Fixed