[FIR] Ditch `KtTypeReference` from `FirDiagnosticsList`
Plugins can generate `FirTypeRef`s. There's no reliable way to provide
a `KtTypeReference` for them, so most of the time they will contain
incompatible ones. This is true for everything that can be generated
from a plugin, but until now we've only really faced it with type refs.

See the history of changes in `entityTypeSuperClass.fir.txt` for an
example.

^KT-68375